Fix displaying orders for a customer

+++ src/main/java/app/entities/orders/Order.java
@@ -52,7 +52,7 @@ public class Order {
         float totalPrice = 0;
 
         for (OrderItem item : orderItems) {
-            totalPrice += (float) (item.getProduct().getSalesPrice() * item.getQuantity());
+            totalPrice += item.getProduct().getSalesPrice() * item.getQuantity();
         }
         return totalPrice;
     }
