+++ src/main/java/app/entities/orders/OrderItem.java
@@ -38,6 +38,6 @@ public class OrderItem {
     }
 
     public float getTotalPrice() {
-        return (float) (product.getSalesPrice() * quantity);
+        return product.getSalesPrice() * quantity;
     }
 }
